backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.1.1
Werkzeug>=3.0.1
SQLAlchemy>=2.0.25
python-dotenv>=1.0.0
requests>=2.31.0
gunicorn>=21.2.0
openai==1.3.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
PyJWT==2.8.0
cryptography==41.0.7
jsonschema==4.19.2
regex==2023.10.3 